+++ hypervisor/include/arch/x86/guest/vcpu.h
@@ -50,6 +50,7 @@
 #ifndef ASSEMBLER
 
 #include <guest.h>
+#include <gpr.h>
 
 enum vcpu_state {
 	VCPU_INIT,
@@ -64,28 +65,6 @@ enum vm_cpu_mode {
 	CPU_MODE_PROTECTED,
 	CPU_MODE_COMPATIBILITY,		/* IA-32E mode (CS.L = 0) */
 	CPU_MODE_64BIT,			/* IA-32E mode (CS.L = 1) */
-};
-
-/* General-purpose register layout aligned with the general-purpose register idx
- * when vmexit, such as vmexit due to CR access, refer to SMD Vol.3C 27-6.
- */
-struct cpu_gp_regs {
-	uint64_t rax;
-	uint64_t rcx;
-	uint64_t rdx;
-	uint64_t rbx;
-	uint64_t rsp;
-	uint64_t rbp;
-	uint64_t rsi;
-	uint64_t rdi;
-	uint64_t r8;
-	uint64_t r9;
-	uint64_t r10;
-	uint64_t r11;
-	uint64_t r12;
-	uint64_t r13;
-	uint64_t r14;
-	uint64_t r15;
 };
 
 struct segment_sel {
